upgraded tests
